// models/enums/event-name.ts
export enum EventName {
  CreateSession = 'create-session',
  CheckCode = 'check-code',
  PlayerConnect = 'player-connect',
  StartQuiz = 'start-quiz',
  GetQuestion = 'get-question',
  GetPlayerCount = 'get-player-count',
  SubmitAnswer = 'submit-answer',
  RevealResults = 'reveal-results',
  NextQuestion = 'next-question',
  HostRejoinSession = 'host-rejoin-session',
  PlayerCount = 'player-count',
  ShowQuestion = 'show-question',
  AnswerCount = 'answer-count',
  QuestionResults = 'question-results',
  PlayerResult = 'player-result',
  QuizFinished = 'quiz-finished',
  SessionPending = 'session-pending',
  SessionEnded = 'session-ended',
  SessionResume = 'session-resume',
}